+++ core/components/romanesco/elements/snippets/06_formulas/f_users/iseditor.snippet.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * isEditor snippet
+ *
+ * Check if user is logged in to the manager.
+ *
+ * @author Mark Hamstra
+ */
+
+if ($modx->user instanceof modUser) {
+    if ($modx->user->hasSessionContext('mgr')) {
+        return true;
+    }
+}
+return false;